# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления солнечными панелями и расчета их мощности",
        "title": "Solar Panels Power Calculator API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/api",
    "paths": {
...
        "/panels": {
            "get": {
                "description": "Возвращает список всех солнечных панелей с возможностью фильтрации по мощности",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SolarPanels"
                ],
                "summary": "Получить список солнечных панелей",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Минимальная мощность для фильтрации",
                        "name": "start_value",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Максимальная мощность для фильтрации",
                        "name": "end_value",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список солнечных панелей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/lab_internal_app_ds.SolarPanel"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры фильтрации",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Панели не найдены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новую солнечную панель. Требуется роль модератора.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SolarPanels"
                ],
                "summary": "Добавить новую солнечную панель (Модератор)",
                "parameters": [
                    {
                        "description": "Данные новой панели",
                        "name": "solarPanel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lab_internal_app_DTO.AddSolarPanel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданная панель",
                        "schema": {
                            "$ref": "#/definitions/lab_internal_app_ds.SolarPanel"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/panels/{id}": {
            "get": {
                "description": "Возвращает детальную информацию о конкретной солнечной панели",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SolarPanels"
                ],
                "summary": "Получить солнечную панель по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID солнечной панели",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные солнечной панели",
                        "schema": {
                            "$ref": "#/definitions/lab_internal_app_ds.SolarPanel"
                        }
                    },
                    "404": {
                        "description": "Панель не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет данные существующей солнечной панели. Требуется роль модератора.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SolarPanels"
                ],
                "summary": "Изменить солнечную панель (Модератор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID солнечной панели",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленные данные панели",
                        "name": "solarPanel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lab_internal_app_DTO.ChangeSolarPanel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная панель",
                        "schema": {
                            "$ref": "#/definitions/lab_internal_app_ds.SolarPanel"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Панель не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавляет солнечную панель в текущую заявку пользователя (корзину)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SolarPanels"
                ],
                "summary": "Добавить панель в корзину",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID солнечной панели",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Панель добавлена в заявку",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Панель уже в заявке или удалена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Панель не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Помечает солнечную панель как удаленную. Требуется роль модератора.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SolarPanels"
                ],
                "summary": "Удалить солнечную панель (Модератор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID солнечной панели",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Панель успешно удалена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Панель не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/panels/{id}/image": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Загружает изображение для солнечной панели. Требуется роль модератора.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SolarPanels"
                ],
                "summary": "Загрузить изображение панели (Модератор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID солнечной панели",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Файл изображения",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Панель с обновленным изображением",
                        "schema": {
                            "$ref": "#/definitions/lab_internal_app_ds.SolarPanel"
                        }
                    },
                    "400": {
                        "description": "Файл не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Панель не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "lab_internal_app_DTO.AddSolarPanel": {
            "type": "object",
            "required": [
                "depth",
                "efficiency",
                "height",
                "power",
                "title",
                "type",
                "width"
            ],
            "properties": {
                "depth": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "efficiency": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "power": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "lab_internal_app_DTO.ChangeSolarPanel": {
            "type": "object",
            "properties": {
                "depth": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "efficiency": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "power": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
...
        "lab_internal_app_ds.SolarPanel": {
            "type": "object",
            "properties": {
                "depth": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "efficiency": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "isDelete": {
                    "type": "boolean"
                },
                "power": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}